Extract in two different files
